+++ tests/test_redis_core.py
@@ -1,9 +1,13 @@
 """Testing the Redis core of cachier."""
 
+import sys
+import time
 import hashlib
 import queue
 import threading
 import warnings
+import contextlib
+import pickle
 from datetime import datetime, timedelta
 from random import random
 from time import sleep
@@ -751,3 +755,513 @@ def test_redis_clear_being_calculated_with_pipeline():
     pipeline_mock.hset.assert_any_call(b"key2", "processing", "false")
     pipeline_mock.hset.assert_any_call(b"key3", "processing", "false")
     pipeline_mock.execute.assert_called_once()
+
+
+# Tes Redis import error handling (lines 14-15)
+@pytest.mark.redis
+def test_redis_import_error_handling():
+    """Test Redis backend when redis package is not available."""
+    # This test is already covered by test_redis_import_warning
+    # but let's ensure the specific lines are hit
+    with patch.dict(sys.modules, {"redis": None}):
+        # Force reload of redis core module
+        if "cachier.cores.redis" in sys.modules:
+            del sys.modules["cachier.cores.redis"]
+
+        # Test import failure
+        try:
+            from cachier.cores.redis import _RedisCore  # noqa: F401
+
+            pytest.skip("Redis is installed, cannot test import error")
+        except ImportError:
+            pass  # Expected behavior
+
+
+# Test Redis corrupted entry handling (lines 112-114)
+@pytest.mark.redis
+def test_redis_corrupted_entry_handling():
+    """Test Redis backend with corrupted cache entries."""
+    import redis
+
+    client = redis.Redis(host="localhost", port=6379, decode_responses=False)
+
+    try:
+        # Test connection
+        client.ping()
+    except redis.ConnectionError:
+        pytest.skip("Redis server not available")
+
+    @cachier(backend="redis", redis_client=client)
+    def test_func(x):
+        return x * 2
+
+    # Clear cache
+    test_func.clear_cache()
+
+    # Manually insert corrupted data
+    cache_key = "cachier:test_coverage_gaps:test_func:somehash"
+    client.hset(cache_key, "value", b"corrupted_pickle_data")
+    client.hset(cache_key, "time", str(time.time()).encode())
+    client.hset(cache_key, "stale", b"0")
+    client.hset(cache_key, "being_calculated", b"0")
+
+    # Try to access - should handle corrupted data gracefully
+    result = test_func(42)
+    assert result == 84
+
+    test_func.clear_cache()
+
+
+# TestRedis deletion failure during eviction (lines 133-135)
+@pytest.mark.redis
+def test_redis_deletion_failure_during_eviction():
+    """Test Redis LRU eviction with deletion failures."""
+    import redis
+
+    client = redis.Redis(host="localhost", port=6379, decode_responses=False)
+
+    try:
+        client.ping()
+    except redis.ConnectionError:
+        pytest.skip("Redis server not available")
+
+    @cachier(
+        backend="redis",
+        redis_client=client,
+        cache_size_limit="100B",  # Very small limit to trigger eviction
+    )
+    def test_func(x):
+        return "x" * 50  # Large result to fill cache quickly
+
+    # Clear cache
+    test_func.clear_cache()
+
+    # Fill cache to trigger eviction
+    test_func(1)
+
+    # Mock delete to fail
+    original_delete = client.delete
+    delete_called = []
+
+    def mock_delete(*args):
+        delete_called.append(args)
+        # Fail on first delete attempt
+        if len(delete_called) == 1:
+            raise redis.RedisError("Mocked deletion failure")
+        return original_delete(*args)
+
+    client.delete = mock_delete
+
+    try:
+        # This should trigger eviction and handle the deletion failure
+        test_func(2)
+        # Verify delete was attempted
+        assert len(delete_called) > 0
+    finally:
+        client.delete = original_delete
+        test_func.clear_cache()
+
+
+# Test Redis non-bytes timestamp handling (line 364)
+@pytest.mark.redis
+def test_redis_non_bytes_timestamp():
+    """Test Redis backend with non-bytes timestamp values."""
+    import redis
+
+    from cachier.cores.redis import _RedisCore
+
+    client = redis.Redis(host="localhost", port=6379, decode_responses=False)
+
+    try:
+        client.ping()
+    except redis.ConnectionError:
+        pytest.skip("Redis server not available")
+
+    @cachier(
+        backend="redis", redis_client=client, stale_after=timedelta(seconds=10)
+    )
+    def test_func(x):
+        return x * 2
+
+    # Clear cache
+    test_func.clear_cache()
+
+    # Create an entry
+    test_func(1)
+
+    # Manually modify timestamp to be a string instead of bytes
+    keys = list(
+        client.scan_iter(match="cachier:test_coverage_gaps:test_func:*")
+    )
+    if keys:
+        # Force timestamp to be a string (non-bytes)
+        client.hset(keys[0], "time", "not_a_number")
+
+    # Create a separate core instance to test stale deletion
+    core = _RedisCore(
+        hash_func=None,
+        redis_client=client,
+        wait_for_calc_timeout=0,
+    )
+    core.set_func(test_func)
+
+    # Try to delete stale entries - should handle non-bytes timestamp
+    # gracefully
+    with contextlib.suppress(Exception):
+        core.delete_stale_entries(timedelta(seconds=1))
+
+    test_func.clear_cache()
+
+
+# Test Redis missing import
+@pytest.mark.redis
+def test_redis_import_error():
+    """Test Redis client initialization warning."""
+    # Test creating a Redis core without providing a client
+    import warnings
+
+    with warnings.catch_warnings(record=True):
+        warnings.simplefilter("always")
+
+        with pytest.raises(Exception, match="redis_client"):
+
+            @cachier(backend="redis", redis_client=None)
+            def test_func():
+                return "test"
+
+
+# Test Redis corrupted entry in LRU eviction
+@pytest.mark.redis
+def test_redis_lru_corrupted_entry():
+    """Test Redis LRU eviction with corrupted entry."""
+    import redis
+
+    client = redis.Redis(host="localhost", port=6379, decode_responses=False)
+    try:
+        client.ping()
+    except redis.ConnectionError:
+        pytest.skip("Redis not available")
+
+    @cachier(
+        backend="redis",
+        redis_client=client,
+        cache_size_limit="200B",  # Small limit
+    )
+    def test_func(x):
+        return f"result_{x}" * 10  # ~60 bytes per entry
+
+    test_func.clear_cache()
+
+    # Add valid entry
+    test_func(1)
+
+    # Add corrupted entry manually
+    from cachier.cores.redis import _RedisCore
+
+    core = _RedisCore(
+        hash_func=None,
+        redis_client=client,
+        wait_for_calc_timeout=0,
+        cache_size_limit="200B",
+    )
+    core.set_func(test_func)
+
+    # Create corrupted entry
+    bad_key = f"{core.key_prefix}:{core._func_str}:badkey"
+    client.hset(bad_key, "value", b"not_valid_pickle")
+    client.hset(bad_key, "time", str(time.time()).encode())
+    client.hset(bad_key, "stale", b"0")
+    client.hset(bad_key, "being_calculated", b"0")
+
+    # This should trigger eviction and handle the corrupted entry
+    test_func(2)
+    test_func(3)
+
+    test_func.clear_cache()
+
+
+# Test Redis deletion failure in eviction
+@pytest.mark.redis
+def test_redis_eviction_delete_failure():
+    """Test Redis eviction handling delete failures."""
+    import warnings
+
+    import redis
+
+    client = redis.Redis(host="localhost", port=6379, decode_responses=False)
+    try:
+        client.ping()
+    except redis.ConnectionError:
+        pytest.skip("Redis not available")
+
+    # Create a unique function to avoid conflicts
+    @cachier(backend="redis", redis_client=client, cache_size_limit="150B")
+    def test_eviction_func(x):
+        return "x" * 50  # Large value
+
+    test_eviction_func.clear_cache()
+
+    # Fill cache to trigger eviction
+    test_eviction_func(100)
+
+    # This should trigger eviction
+    with warnings.catch_warnings(record=True):
+        # Ignore warnings about eviction failures
+        warnings.simplefilter("always")
+        test_eviction_func(200)
+
+    # Verify both values work (even if eviction had issues)
+    result1 = test_eviction_func(100)
+    result2 = test_eviction_func(200)
+
+    assert result1 == "x" * 50
+    assert result2 == "x" * 50
+
+    test_eviction_func.clear_cache()
+
+
+# Test Redis stale deletion with size tracking
+@pytest.mark.redis
+def test_redis_stale_delete_size_tracking():
+    """Test Redis stale deletion updates cache size."""
+    import redis
+
+    client = redis.Redis(host="localhost", port=6379, decode_responses=False)
+    try:
+        client.ping()
+    except redis.ConnectionError:
+        pytest.skip("Redis not available")
+
+    @cachier(
+        backend="redis",
+        redis_client=client,
+        cache_size_limit="1KB",
+        stale_after=timedelta(seconds=0.1),
+    )
+    def test_func(x):
+        return "data" * 20
+
+    test_func.clear_cache()
+
+    # Create entries
+    test_func(1)
+    test_func(2)
+
+    # Wait for staleness
+    sleep(0.2)
+
+    # Get the core
+    from cachier.cores.redis import _RedisCore
+
+    core = _RedisCore(
+        hash_func=None,
+        redis_client=client,
+        wait_for_calc_timeout=0,
+        cache_size_limit="1KB",
+    )
+    core.set_func(test_func)
+
+    # Delete stale entries - this should update cache size
+    core.delete_stale_entries(timedelta(seconds=0.1))
+
+    # Verify size tracking by adding new entry
+    test_func(3)
+
+    test_func.clear_cache()
+
+
+@pytest.mark.redis
+def test_redis_lru_eviction_edge_cases():
+    """Test Redis LRU eviction edge cases for coverage."""
+    from cachier.cores.redis import _RedisCore
+
+    redis_client = _test_redis_getter()
+
+    # Test 1: Corrupted data during LRU eviction (lines 112-114)
+    core = _RedisCore(
+        hash_func=None, redis_client=redis_client, cache_size_limit=100
+    )
+
+    def mock_func(x):
+        return x * 2
+
+    core.set_func(mock_func)
+
+    # Add entries with corrupted metadata
+    for i in range(3):
+        key = core._get_redis_key(f"key{i}")
+        redis_client.hset(key, "value", pickle.dumps(i * 2))
+        redis_client.hset(
+            key, "time", pickle.dumps(datetime.now().timestamp())
+        )
+        if i == 1:
+            # Corrupt metadata for one entry
+            redis_client.hset(key, "last_access", "invalid_json")
+            redis_client.hset(key, "size", "not_a_number")
+        else:
+            redis_client.hset(key, "last_access", str(time.time()))
+            redis_client.hset(key, "size", "20")
+
+    # Set high cache size to trigger eviction
+    redis_client.set(core._cache_size_key, "1000")
+
+    # Should handle corrupted entries gracefully
+    core._evict_lru_entries(redis_client, 1000)
+
+    # Test 2: No eviction needed (line 138)
+    # Clear and set very low cache size
+    pattern = f"{core.key_prefix}:{core._func_str}:*"
+    for key in redis_client.scan_iter(match=pattern):
+        if b"__size__" not in key:
+            redis_client.delete(key)
+
+    redis_client.set(core._cache_size_key, "10")
+    # Should not evict anything
+    core._evict_lru_entries(threshold_fraction=0.9)
+
+
+@pytest.mark.redis
+def test_redis_clear_and_delete_edge_cases():
+    """Test Redis clear and delete operations edge cases."""
+    from cachier.cores.redis import _RedisCore
+
+    redis_client = _test_redis_getter()
+
+    # Test 1: clear_being_calculated with no keys (line 325)
+    core = _RedisCore(hash_func=None, redis_client=redis_client)
+
+    def mock_func():
+        pass
+
+    core.set_func(mock_func)
+
+    # Ensure no keys exist
+    pattern = f"{core.key_prefix}:{core._func_str}:*"
+    for key in redis_client.scan_iter(match=pattern):
+        redis_client.delete(key)
+
+    # Should handle empty key set gracefully
+    core.clear_being_calculated()
+
+    # Test 2: delete_stale_entries with special keys (line 352)
+    core2 = _RedisCore(hash_func=None, redis_client=redis_client)
+    core2.stale_after = timedelta(seconds=1)
+
+    def mock_func2():
+        pass
+
+    core2.set_func(mock_func2)
+
+    # Add stale entries
+    for i in range(2):
+        key = core2._get_redis_key(f"entry{i}")
+        redis_client.hset(key, "value", pickle.dumps(f"value{i}"))
+        redis_client.hset(
+            key,
+            "time",
+            pickle.dumps((datetime.now() - timedelta(seconds=2)).timestamp()),
+        )
+
+    # Add special cache size key
+    redis_client.set(core2._cache_size_key, "100")
+
+    # Delete stale - should skip special keys
+    core2.delete_stale_entries(timedelta(seconds=1))
+
+    # Special key should still exist
+    assert redis_client.exists(core2._cache_size_key)
+
+    # Test 3: Non-bytes timestamp (line 364)
+    key = core2._get_redis_key("nonbytes")
+    redis_client.hset(key, "value", pickle.dumps("test"))
+    # String timestamp instead of bytes
+    redis_client.hset(
+        key, "time", str((datetime.now() - timedelta(seconds=2)).timestamp())
+    )
+
+    core2.delete_stale_entries(timedelta(seconds=1))
+    # Should handle string timestamp
+    assert not redis_client.exists(key)
+
+
+@pytest.mark.redis
+def test_redis_delete_stale_size_handling():
+    """Test Redis delete_stale_entries size handling."""
+    from cachier.cores.redis import _RedisCore
+
+    redis_client = _test_redis_getter()
+
+    # Test 1: Corrupted size data (lines 374-375)
+    core = _RedisCore(
+        hash_func=None, redis_client=redis_client, cache_size_limit=1000
+    )
+    core.stale_after = timedelta(seconds=1)
+
+    def mock_func():
+        pass
+
+    core.set_func(mock_func)
+
+    # Add entries with one having corrupted size
+    for i in range(3):
+        key = core._get_redis_key(f"item{i}")
+        value = pickle.dumps(f"result{i}")
+        redis_client.hset(key, "value", value)
+        redis_client.hset(
+            key,
+            "time",
+            pickle.dumps((datetime.now() - timedelta(seconds=2)).timestamp()),
+        )
+        if i == 1:
+            redis_client.hset(key, "size", "invalid_size")
+        else:
+            redis_client.hset(key, "size", str(len(value)))
+
+    # Should handle corrupted size gracefully
+    core.delete_stale_entries(timedelta(seconds=1))
+
+    # Test 2: No cache_size_limit (line 380)
+    core2 = _RedisCore(hash_func=None, redis_client=redis_client)
+    core2.stale_after = timedelta(seconds=1)
+    core2.cache_size_limit = None
+
+    def mock_func2():
+        pass
+
+    core2.set_func(mock_func2)
+
+    # Add stale entries
+    for i in range(2):
+        key = core2._get_redis_key(f"old{i}")
+        redis_client.hset(key, "value", pickle.dumps(f"old{i}"))
+        redis_client.hset(
+            key,
+            "time",
+            pickle.dumps((datetime.now() - timedelta(seconds=2)).timestamp()),
+        )
+        redis_client.hset(key, "size", "50")
+
+    core2.delete_stale_entries(timedelta(seconds=1))
+
+    # Test 3: Nothing to delete (line 380)
+    core3 = _RedisCore(
+        hash_func=None, redis_client=redis_client, cache_size_limit=1000
+    )
+    core3.stale_after = timedelta(days=1)
+
+    def mock_func3():
+        pass
+
+    core3.set_func(mock_func3)
+
+    # Add fresh entries
+    for i in range(2):
+        key = core3._get_redis_key(f"fresh{i}")
+        redis_client.hset(key, "value", pickle.dumps(f"fresh{i}"))
+        redis_client.hset(
+            key, "time", pickle.dumps(datetime.now().timestamp())
+        )
+        redis_client.hset(key, "size", "30")
+
+    # Nothing should be deleted
+    core3.delete_stale_entries(timedelta(days=1))
